+++ variants/tests/factories.py
@@ -314,7 +314,7 @@ class CaseFactory(factory.django.DjangoModelFactory):
 
     @factory.lazy_attribute_sequence
     def pedigree(self, n):
-        if self.structure not in ("singleton", "trio", "quartet", "quintet"):
+        if self.structure not in ("singleton", "duo", "trio", "quartet", "quintet"):
             raise ValueError("Invalid structure type!")
         elif self.structure == "singleton":
             return [
@@ -326,6 +326,26 @@ class CaseFactory(factory.django.DjangoModelFactory):
                     "affected": 2,
                     "has_gt_entries": True,
                 }
+            ]
+        elif self.structure == "duo":
+            father = "father_%03d-N1-DNA1-WES1" % n
+            return [
+                {
+                    "patient": self.index,
+                    "father": father,
+                    "mother": "0",
+                    "sex": self.sex,
+                    "affected": 2,  # always affected
+                    "has_gt_entries": True,
+                },
+                {
+                    "patient": father,
+                    "father": "0",
+                    "mother": "0",
+                    "sex": 1,  # always male
+                    "affected": 2 if self.inheritance == "dominant" else 1,
+                    "has_gt_entries": True,
+                },
             ]
         elif self.structure == "trio":
             # Father and mother name